road_hazard_reporter/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
